+++ readScoress.py
