+++ readTimess.py
